move functions supporting rmi command to libkpod/image
