+++ cmd/kpod/rmi.go
@@ -3,10 +3,6 @@ package main
 import (
 	"fmt"
 
-	is "github.com/containers/image/storage"
-	"github.com/containers/image/transports"
-	"github.com/containers/image/transports/alltransports"
-	"github.com/containers/image/types"
 	"github.com/containers/storage"
 	libkpodimage "github.com/kubernetes-incubator/cri-o/libkpod/image"
 	"github.com/pkg/errors"
@@ -49,7 +45,7 @@ func rmiCmd(c *cli.Context) error {
 	}
 
 	for _, id := range args {
-		image, err := getImage(id, store)
+		image, err := libkpodimage.GetImage(store, id)
 		if err != nil {
 			return errors.Wrapf(err, "could not get image %q", id)
 		}
@@ -75,7 +71,7 @@ func rmiCmd(c *cli.Context) error {
 				// If it is forced, we have to untag the image so that it can be deleted
 				image.Names = image.Names[:0]
 			} else {
-				name, err2 := untagImage(id, image, store)
+				name, err2 := libkpodimage.UntagImage(store, image, id)
 				if err2 != nil {
 					return err
 				}
@@ -85,7 +81,7 @@ func rmiCmd(c *cli.Context) error {
 			if len(image.Names) > 0 {
 				continue
 			}
-			id, err := removeImage(image, store)
+			id, err := libkpodimage.RemoveImage(image, store)
 			if err != nil {
 				return err
 			}
@@ -96,69 +92,8 @@ func rmiCmd(c *cli.Context) error {
 	return nil
 }
 
-func getImage(id string, store storage.Store) (*storage.Image, error) {
-	var ref types.ImageReference
-	ref, err := properImageRef(id)
-	if err != nil {
-		//logrus.Debug(err)
-	}
-	if ref == nil {
-		if ref, err = storageImageRef(store, id); err != nil {
-			//logrus.Debug(err)
-		}
-	}
-	if ref == nil {
-		if ref, err = storageImageID(store, id); err != nil {
-			//logrus.Debug(err)
-		}
-	}
-	if ref != nil {
-		image, err2 := is.Transport.GetStoreImage(store, ref)
-		if err2 != nil {
-			return nil, err2
-		}
-		return image, nil
-	}
-	return nil, err
-}
-
-func untagImage(imgArg string, image *storage.Image, store storage.Store) (string, error) {
-	// Remove name from image.Names and set the new name in the ImageStore
-	imgStore, err := store.ImageStore()
-	if err != nil {
-		return "", errors.Wrap(err, "could not untag image")
-	}
-	newNames := []string{}
-	removedName := ""
-	for _, name := range image.Names {
-		if libkpodimage.MatchesReference(name, imgArg) {
-			removedName = name
-			continue
-		}
-		newNames = append(newNames, name)
-	}
-	imgStore.SetNames(image.ID, newNames)
-	err = imgStore.Save()
-	return removedName, err
-}
-
-func removeImage(image *storage.Image, store storage.Store) (string, error) {
-	imgStore, err := store.ImageStore()
-	if err != nil {
-		return "", errors.Wrapf(err, "could not open image store")
-	}
-	err = imgStore.Delete(image.ID)
-	if err != nil {
-		return "", errors.Wrapf(err, "could not remove image")
-	}
-	err = imgStore.Save()
-	if err != nil {
-		return "", errors.Wrapf(err, "could not save image store")
-	}
-	return image.ID, nil
-}
-
 // Returns a list of running containers associated with the given ImageReference
+// TODO: replace this with something in libkpod
 func runningContainers(image *storage.Image, store storage.Store) ([]string, error) {
 	ctrIDs := []string{}
 	ctrStore, err := store.ContainerStore()
@@ -178,6 +113,7 @@ func runningContainers(image *storage.Image, store storage.Store) ([]string, err
 	return ctrIDs, nil
 }
 
+// TODO: replace this with something in libkpod
 func removeContainers(ctrIDs []string, store storage.Store) error {
 	ctrStore, err := store.ContainerStore()
 	if err != nil {
@@ -189,51 +125,4 @@ func removeContainers(ctrIDs []string, store storage.Store) error {
 		}
 	}
 	return nil
-}
-
-// If it's looks like a proper image reference, parse it and check if it
-// corresponds to an image that actually exists.
-func properImageRef(id string) (types.ImageReference, error) {
-	var ref types.ImageReference
-	var err error
-	if ref, err = alltransports.ParseImageName(id); err == nil {
-		if img, err2 := ref.NewImage(nil); err2 == nil {
-			img.Close()
-			return ref, nil
-		}
-		return nil, fmt.Errorf("error confirming presence of image reference %q: %v", transports.ImageName(ref), err)
-	}
-	return nil, fmt.Errorf("error parsing %q as an image reference: %v", id, err)
-}
-
-// If it's looks like an image reference that's relative to our storage, parse
-// it and check if it corresponds to an image that actually exists.
-func storageImageRef(store storage.Store, id string) (types.ImageReference, error) {
-	var ref types.ImageReference
-	var err error
-	if ref, err = is.Transport.ParseStoreReference(store, id); err == nil {
-		if img, err2 := ref.NewImage(nil); err2 == nil {
-			img.Close()
-			return ref, nil
-		}
-		return nil, fmt.Errorf("error confirming presence of storage image reference %q: %v", transports.ImageName(ref), err)
-	}
-	return nil, fmt.Errorf("error parsing %q as a storage image reference: %v", id, err)
-}
-
-// If it might be an ID that's relative to our storage, parse it and check if it
-// corresponds to an image that actually exists.  This _should_ be redundant,
-// since we already tried deleting the image using the ID directly above, but it
-// can't hurt either.
-func storageImageID(store storage.Store, id string) (types.ImageReference, error) {
-	var ref types.ImageReference
-	var err error
-	if ref, err = is.Transport.ParseStoreReference(store, "@"+id); err == nil {
-		if img, err2 := ref.NewImage(nil); err2 == nil {
-			img.Close()
-			return ref, nil
-		}
-		return nil, fmt.Errorf("error confirming presence of storage image reference %q: %v", transports.ImageName(ref), err)
-	}
-	return nil, fmt.Errorf("error parsing %q as a storage image reference: %v", "@"+id, err)
 }
